% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_r_to_sql_df.R
\name{gta_r_to_sql_df}
\alias{gta_r_to_sql_df}
\title{A little function that translates the GTA's naming rules into SQL. Here for data frame (names and column names).}
\usage{
gta_r_to_sql_df(convert.df = NULL)
}
\arguments{
\item{convert.df}{Name of the data frame you want to convert (in quotes).}
}
\description{
Returns renamed dataframe.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
