% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_r_to_sql_var.R
\name{gta_r_to_sql_var}
\alias{gta_r_to_sql_var}
\title{A little function that translates the GTA's naming rules into SQL. Here for variable/vector object.}
\usage{
gta_r_to_sql_var(convert.var = NULL)
}
\arguments{
\item{convert.var}{Name of the vector you want to convert (in quotes).}
}
\description{
Returns renamed vector.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
