% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_coltype_initialise.R
\name{gta_sql_coltype_initialise}
\alias{gta_sql_coltype_initialise}
\title{Create a table that collects the classes of Rstudio's data frames when you upload them.}
\usage{
gta_sql_coltype_initialise(db.connection = "pool")
}
\arguments{
\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
The columns will be automatically converted back into those classes when you download them from the SQL database.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
