% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_create_table.R
\name{gta_sql_create_table}
\alias{gta_sql_create_table}
\title{Create a new table in your SQL database.}
\usage{
gta_sql_create_table(write.df = NULL, create.primary.key = NULL,
  create.primary.auto.incr = F, create.foreign.key = NULL,
  create.foreign.key.parent = NULL, create.foreign.key.del.cascade = T,
  append.existing = T, db.connection = "pool")
}
\arguments{
\item{write.df}{The name of the data frame (in quotes) you want to add.}

\item{create.primary.key}{Name of the column that is the primary key. Can only be one. GTA or SQL naming is fine.}

\item{create.primary.auto.incr}{Do you want the primary key to be auto incrementing? Default is TRUE.}

\item{create.foreign.key}{Vector with names of columns that are foreign keys. Can use GTA or SQL naming.}

\item{create.foreign.key.parent}{On what table is the foreign key the primary key?}

\item{create.foreign.key.del.cascade}{What happens to the lines of foreign key if a value is deleted from the primary key is deleted. Default is that they are removed as well.}

\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
Creates a data table with the name of your dataframe. Uploads the data frame thereafter.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
