% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_get_value.R
\name{gta_sql_get_value}
\alias{gta_sql_get_value}
\title{Get a value from the SQL database.}
\usage{
gta_sql_get_value(query = NULL, db.connection = "pool")
}
\arguments{
\item{query}{Your query in SQL language (incl SQL naming!).}

\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
Just a wrapper, really. Returns a scalar or a data frame, depending on how many results you have.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
