% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_keytype_initialise.R
\name{gta_sql_keytype_initialise}
\alias{gta_sql_keytype_initialise}
\title{Create a table that collects the keytypes}
\usage{
gta_sql_keytype_initialise(db.connection = "pool")
}
\arguments{
\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
Create a table that collects the keytypes
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
