% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_keytype_record.R
\name{gta_sql_keytype_record}
\alias{gta_sql_keytype_record}
\title{Record the key in the data frame your uploading}
\usage{
gta_sql_keytype_record(record.frame = NULL, record.column = NULL,
  record.type = NULL, record.style = NULL, db.connection = "pool")
}
\arguments{
\item{record.frame}{The name of the data frame (in quotes) whose column classes you want to record.}

\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
The columns will be automatically converted back into those classes when you download them from the SQL database.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
