% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_library_update.R
\name{gta_sql_library_update}
\alias{gta_sql_library_update}
\title{Get the latest version from GitHub.}
\usage{
gta_sql_library_update(x)
}
\value{
Be up to date with our latest functions.
}
\description{
Syncs your local library with our latest GTA GitHub release.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
