% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_load_table.R
\name{gta_sql_load_table}
\alias{gta_sql_load_table}
\title{Load a table from the SQL database into your local environment.}
\usage{
gta_sql_load_table(load.table = NULL, db.connection = "pool")
}
\arguments{
\item{load.table}{The name of the table you want to load. SQL or GTA naming is fine.}

\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
Load a table from the SQL database into your local environment.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
