% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_pool_open.R
\name{gta_sql_pool_open}
\alias{gta_sql_pool_open}
\title{Record the classes of Rstudio's data frames when you upload them.}
\usage{
gta_sql_pool_open(db.title = NULL, db.name = NULL, db.user = NULL,
  db.password = NULL, db.host = NULL)
}
\arguments{
\item{db.title}{Name of the connection in the GTA's key chain (or as recorded in this function).}

\item{db.name}{connection info.}

\item{db.user}{connection info.}

\item{db.password}{connection info.}

\item{db.host}{connection info.}

\item{db.name}{connection info.}
}
\description{
Note that the default connects you to ricardo-dev. Add “ricardo” as db.title’s value if you want to connect to the production database.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
