% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_set_table_keys.R
\name{gta_sql_set_table_keys}
\alias{gta_sql_set_table_keys}
\title{A wrapper function that sets the key types in your SQL table.}
\usage{
gta_sql_set_table_keys(table.name = NULL, primary.key = NULL,
  primary.auto.incr = F, foreign.key = NULL,
  foreign.key.parent = NULL, foreign.key.del.cascade = T,
  db.connection = "pool")
}
\arguments{
\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
A wrapper function that sets the key types in your SQL table.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
