% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gta_sql_update_table.R
\name{gta_sql_update_table}
\alias{gta_sql_update_table}
\title{Update a specific portion of an existing table in your SQL database}
\usage{
gta_sql_update_table(query = NULL, db.connection = "pool")
}
\arguments{
\item{query}{A string with the SQL query you want to execute.}

\item{db.connection}{Specify the database connection you want to use. Default is 'pool'.}
}
\description{
Just a wrapper to make the connection easier. You have to type out the query in SQL, no help here.
}
\references{
www.globaltradealert.org
}
\author{
Global Trade Alert
}
